Did major refactoring of the TOC and logging handling code plus more
The whole process of handling TOC elements and their properties was refactored (i.e. still is) and reads TOC contents for Parameters (port 2) and Logging (port 5) correctly. The development of moving sensor values over to the controller part and the user program is right in the middle at the moment and under active development. The main cycle() function in the CCrazyflie class now features several states for initialization and normal operation. This way, no unwanted (undefined) control messages get sent over the communication channels before everything is set up.

Two new example programs were added - one for actual flying (although it does not let the copter fly at the moment - this is subject to further development as it will take sensor readings into account) and one for pure dev purposes on my side (feel free to check it out, though).

The sensor parts will be completed somewhere during the next few days, depending on when I find time for it. I'm pretty confident so far that the lib is going to nicely serve it's purpose as the IMU already gives out very nice readings that are recorded in the lib.