Added handling of all variable types in log data element access
Also, added a friendly warning in the comments for FP16. This type has been implemented, but I have no idea if it works the way I think it does. I'm 75% sure this is working, but will check back once I find time for it. Be careful when using it at the moment.