Use throw when something bad happens when dealing with log values
I've thought a while if it's a good idea to use throw when an error
occured, but if it happens we most likely already have lost control
over the Crazyflie. So there is no need for additional catches and program
flow will become more simple.

While beeing there remove unused checkAndSetLogValue() helpers.